requirements.txt
Flask
Flask-Cors
requests
rq
cachecontrol
python-jose[cryptography]
gunicorn~=20.0.4
gevent~=1.4.0
